refactor: replace syncIssueMetaData with updateDirectoryIssue and update parameter names for clarity
