+++ src/twitter/initialize-twitter-map.ts
@@ -1,4 +1,3 @@
-import { readFile } from "fs/promises";
 import { commitTwitterMap } from "../git";
 
 export type TwitterMap = Record<string, string>;
@@ -6,9 +5,13 @@ export type TwitterMap = Record<string, string>;
 export async function initializeTwitterMap() {
   let twitterMap: TwitterMap = {};
   try {
-    twitterMap = JSON.parse(await readFile("./twitter-map.json", "utf8"));
+    const response = await fetch("https://raw.githubusercontent.com/0x4007/devpool-directory/__STORAGE__/twitter-map.json");
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+    twitterMap = await response.json();
   } catch (error) {
-    console.log("Couldn't find twitter map artifact, creating a new one");
+    console.log("Couldn't fetch twitter map, creating a new one");
     await commitTwitterMap(twitterMap);
   }
   return twitterMap;
